asynxd/requirements.txt
anyjson>=0.3.3
asynx_core>=0.4dev
daemonize>=2.3
flask>=0.10.1
Flask-And-Redis>=0.5
Flask-Script>=2.0.5
gevent>=1.0.1
gunicorn>=18.0
python-dateutil>=2.2
pytz>2014
sh>=1.08
simplejson>=3.5.2
voluptuous>=0.8.5
